+++ projeto-API-DB/para-sala/livros/server.js
@@ -1,6 +1,6 @@
 const app  = require ("./src/app.js");
 
-const port = 3000
+const port = 8080
 
 app.listen(port, () => {
     console.log(`Servidor está rodando na porta ${port}`)
