+++ projeto-API-DB/para-sala/livros/src/app.js
@@ -3,8 +3,10 @@ const index = require ("./routes/index.js");
 const livros = require ("./routes/livros.js");
 const db = require ("./config/dbConnect.js");
 
+//quando a conexão pode dar erro
 db.on('error', console.log.bind (console, 'Erro de conexão'))
 
+//quando dá certo
 db.once('open', () => {
     console.log ('conexão com o banco feita com sucesso')
 });
@@ -14,7 +16,8 @@ const app = express();
 
 app.use(express.json());
 
-// transmissão de comunicação http x client
+// cabeçalho das informações que enviamos
+
 app.use(function (req, res, next) {
     res.header("Access-Control-Allow-Origin", "*") 
     res.header(
